+++ src/cola_layout.js
@@ -8,6 +8,7 @@
 dc_graph.cola_layout = function(id) {
     var _layoutId = id || uuid();
     var _d3cola = null;
+    var _setcola_nodes;
     var _dispatch = d3.dispatch('tick', 'start', 'end');
     var _flowLayout;
     // node and edge objects shared with cola.js, preserved from one iteration
@@ -108,7 +109,8 @@ dc_graph.cola_layout = function(id) {
                 wnodes,
                 wedges.map(function(e) {
                     return {dcg_edgeKey: e.dcg_edgeKey};
-                })
+                }),
+                _setcola_nodes
             );
         }
         _d3cola.on('tick', /* _tick = */ function() {
@@ -128,6 +130,7 @@ dc_graph.cola_layout = function(id) {
                 .gap(10) //default value is 10, can be customized in setcolaSpec
                 .layout();
 
+            _setcola_nodes = setcola_result.nodes.filter(function(n) { return n._cid; });
             _d3cola.nodes(setcola_result.nodes)
                 .links(setcola_result.links)
                 .constraints(setcola_result.constraints)
@@ -192,7 +195,7 @@ dc_graph.cola_layout = function(id) {
         },
         optionNames: function() {
             return ['handleDisconnected', 'lengthStrategy', 'baseLength', 'flowLayout',
-                    'tickSize', 'groupConnected', 'setcolaSpec']
+                    'tickSize', 'groupConnected', 'setcolaSpec', 'setcolaNodes']
                 .concat(graphviz_keys);
         },
         passThru: function() {
@@ -282,8 +285,14 @@ dc_graph.cola_layout = function(id) {
         tickSize: property(1),
         groupConnected: property(false),
         setcolaSpec: property(null),
+        setcolaNodes: function() {
+            return _setcola_nodes;
+        },
         extractNodeAttrs: property({}), // {attr: function(node)}
-        extractEdgeAttrs: property({})
+        extractEdgeAttrs: property({}),
+        processExtraWorkerResults: function(setcolaNodes) {
+            _setcola_nodes = setcolaNodes;
+        }
     });
     return engine;
 };